requirements.txt
python-dotenv~=1.0.1
discord~=2.3.2
fastapi~=0.111.0
pyudorandom~=1.0.0